requirements.txt
requests
beautifulsoup4
pandas
schedule
apscheduler
cryptography
pytest
streamlit
xlsxwriter
pyinstaller
flask
fake-useragent
matplotlib
joblib
scikit-learn
psycopg2
sqlalchemy
XGBoost
tensorflow
optuna
ShopifyAPI
requests
selenium
webdriver-manager
boto3
python-amazon-paapi
